backend/requirements.txt
flask
firebase-admin
google-cloud-vision
gunicorn
requests  